pyproject.toml
[tool.poetry]
name = "seoman"
version = "0.1.0"
description = "A Command Line tool that is designed for technical SEOs"
authors = ["seo.do <yagizcan@zeo.org>"]
maintainers = ["Yagiz Degirmenci <yagizcan@zeo.org>"]
keywords = ["seo", "technical seo"]
homepage = "https://seo.do/"
license= "Apache 2.0"
repository = "https://github.com/zeoagency/seoman"

packages = [
    { include = "seoman" },
    { include = "seoman/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.6"
typer = "^0.3.2"
google-api-python-client = "^1.12.5"
google-auth-oauthlib = "^0.4.1"
pytablewriter = "^0.58.0"
requests = "^2.24.0"
halo = "^0.0.30"
pyexcelerate = "^0.9.0"
inquirer = "^2.7.0"
toml = "^0.10.1"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
isort = "^5.6.4"
mypy = "^0.790"

[tool.poetry.scripts]
seoman = 'seoman.main:app'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
